+++ tests/unit/test_mini_frequencies.py
@@ -4,9 +4,7 @@ Test assignment of mini frequencies to the circuit MVD3.
 from pathlib import Path
 
 import pandas as pd
-import pytest
-from brainbuilder.app import cells
-from brainbuilder.exceptions import BrainBuilderError
+from brainbuilder.app.cells import load_mini_frequencies, _assign_mini_frequencies
 
 DATA_PATH = Path(Path(__file__).parent, "data")
 
@@ -20,7 +18,7 @@ def test_mini_frequencies_input():
 
     and the layer information should be in the index.
     """
-    mini_freqs = cells.load_mini_frequencies(DATA_PATH / "mini_frequencies.tsv")
+    mini_freqs = load_mini_frequencies(DATA_PATH / "mini_frequencies.tsv")
     assert mini_freqs.index.name == "layer"
     assert "exc_mini_frequency" in mini_freqs.columns
     assert "inh_mini_frequency" in mini_freqs.columns
@@ -30,32 +28,21 @@ def test_mini_frequencies_assignment():
     """
     Mini frequencies must be assigned to cells.
     """
-    cells_df = pd.read_csv(
-        DATA_PATH / "pre_mini_frequency_assignment_cells.tsv",
-        sep="\t",
-        index_col="gid")
+    cells = pd.DataFrame(data={
+        "subregion": ["plf", "IF", "im", "6b"],
+        })
 
-    mini_freqs = cells.load_mini_frequencies(DATA_PATH / "mini_frequencies.tsv")
+    mini_freqs = pd.DataFrame(data={
+        "exc_mini_frequency": [0.63, 0.26, 0.122, 0.072],
+        "inh_mini_frequency": [0.012, 0.012, 0.012, 0.012],
+        },
+        index=["plf", "IF", "im", "6b"])
+    _assign_mini_frequencies(cells, mini_freqs)
 
-    cells._assign_mini_frequencies(cells_df, mini_freqs)
+    expected = pd.DataFrame(data={
+        "subregion": ["plf", "IF", "im", "6b"],
+        "exc_mini_frequency": [0.63, 0.26, 0.122, 0.072],
+        "inh_mini_frequency": [0.012, 0.012, 0.012, 0.012],
+        })
 
-    expected_df = pd.read_csv(
-        DATA_PATH / "expected-cells-with-mini-frequencies.tsv",
-        sep="\t",
-        index_col="gid")
-
-    pd.testing.assert_frame_equal(cells_df, expected_df)
-
-def test_mini_frequencies_assignment_no_layer():
-    """
-    Mini frequencies cannot be assigned to cells with no layer.
-    """
-    cells_df = pd.read_csv(
-        DATA_PATH / "pre_mini_frequency_assignment_cells_no_layer.tsv",
-        sep="\t",
-        index_col="gid")
-
-    mini_freqs = cells.load_mini_frequencies(DATA_PATH / "mini_frequencies.tsv")
-
-    with pytest.raises(BrainBuilderError, match="Cannot add mini frequencies to cells with no layer"):
-        cells._assign_mini_frequencies(cells_df, mini_freqs)
+    pd.testing.assert_frame_equal(cells, expected)