+++ CHANGELOG.rst
@@ -3,6 +3,7 @@ Changelog
 
 ## 0.18.1
   * `layer` made an optional trait in `brainbuilder.app.cells.place`
+  * `subregion` added as cells property in `brainbuilder.app.cells.place`
 
 ## 0.18.0
   * Add simple circuit splitting for SONATA (BRBLD-90).